Genetic Programming algorithm

Mutation, crossover and GP implementation.

Types and functions

GP_NLS.GP — Method

GP With depth and number of nodes control. The recommended startup is PTC2, but we have the others as well (however, the other methods are based in the koza GP and do not follow restrictions on the maximum number of nodes). To use canonic GP, just disable lm_optimization and choose one of ["ramped", "grow", "full"] initializations. To use GP-NLS, turn on lm_optimization and use "PTC2" as initialization method.

GP(
    X::Matrix{Float64}, 
    y::Vector{Float64},
    fSet::Vector{Func},
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
    minDepth::Int64        = 1,
    maxDepth::Int64        = 5,
    maxSize::Int64         = 25,
    popSize::Int64         = 50,
    gens::Int64            = 50,
    mutationRate::Float64  = 0.25,
    elitism::Bool          = false,
    verbose::Bool          = false,
    init_method::String    = "PTC2", #["ramped", "grow", "full", "PTC2"]
    lm_optimization        = false, 
    keep_linear_transf_box = false
)::AbstractNode

GP_NLS.crossover — Method

Crossover function that does a recombination of the two parents passed as an argument, finding a breakpoint in each of the parent trees and swapping the subtree between those breakpoints. It doesn't change the parents. This crossover controls the number of nodes (not the depth) of the tree, avoiding to exceed the maximum value.

crossover(
    fst_parent::AbstractNode,
    snd_parent::AbstractNode,
    maxDepth::Int64,
    maxSize::Int64)::AbstractNode

GP_NLS.mutation! — Method

Function that implements a traditional substitution mutation in a tree, respecting the maximum past depth. Modifies the tree passed.

mutation!(
    node::AbstractNode,
    maxDepth::Int64,
    maxSize::Int64,
    fSet::Vector{Func},
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
    mutationRate::Float64)::AbstractNode

The mutationRate mutation rate should vary by $[0, 1]$ and determines the chance of occurring a mutation (replacing a random point with a new subtree). If no mutation is performed, then the given node is returned.

GP_NLS.tourn_selection — Method

Function that receives two individuals and makes a simple tournament selection. An individual is a (fitness::Float64, node::AbstractNode) tuple with the tree and it fitness, to avoid recalculations all the time.

tourn_selection(
    ind1::Tuple{Float64, AbstractNode},
    ind2::Tuple{Float64, AbstractNode})::Tuple{AbstractNode, Float64}

The return type is a tuple with the winning individual(that is, the individual winner tuple is returned).