Population Initialization

Implementation of different functions to initialize individual trees, as well as functions to create an entire population.

Types and functions

GP_NLS.PTC2 — Method

Tree creation with the Probabilistic Tree Creator 2 (PTC2) method, described in Two Fast Tree-Creation Algorithms for Genetic Programming, by Sean Luke.

This method looks like Koza's full method, but in addition to respecting a limit of maxDepth depth, it also respects a limit of number of nodes expctdSize.

PTC2 ensures that the depth does not exceed the maximum (in our case, weighted variables count as depth 1), and ensures that the number of nodes does not exceed the expected value added to the highest arity between the functions, that is, $expctdSize + max(arity(f)), f in fSet$.

PTC2(
    fSet::Vector{Func},
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
    maxDepth::Int64,
    expctdSize::Int64)::AbstractNode

Here we adopt that the chance to select a $t$ terminal will be uniform for all possible terminals, and the chance to select a function will also follow the same logic.

GP_NLS._create_random_terminal — Method

Function that receives the set of terminal contents and creates a random terminal node.

_create_random_terminal(
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}})::TerminalNode

Creating a terminal node involves an additional verification step for the case of ERC, which must be replaced with a constant within the range specified.

GP_NLS.full — Method

Function that creates a tree using the full method, inspired by Koza's original work. Receives a set of fSet::Vector{Func} functions, a set of tSet::Vector{Union{Var, WeightedVar, Const, ERC}} terminals, and a maximum depth of maxDepth::Int64 allowed.

Returns any tree with maximum depth maxDepth and using the contents of past functions and terminals.

full(fSet::Vector{Func}, tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
        maxDepth::Int64)::AbstractNode

GP_NLS.grow — Method

Function that creates a tree using the grow method, inspired by Koza's original work. Receives a set of fSet::Vector{Func} functions that will be used in the internal nodes, a set of tSet::Vector{Union{Var, WeightedVar, Const, ERC}} terminals, and a maximum depth of maxDepth::Int64 allowed.

Returns any tree with maximum depth maxDepth created using the functions and terminal sets.

 grow(fSet::Vector{Func}, tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
        maxDepth::Int64)::AbstractNode

Note that there is no minimum size, meaning that a single-node tree can be returned. The maximum depth considers weighted variables as a single node.

GP_NLS.init_pop_PTC2 — Method

Function that initializes a population of size popSize using the PTC2 method.

init_pop_PTC2(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Every initialization functions take the same parameters, but not all do of them makes use of all parameters. This is just to unify the call of initialization functions.

GP_NLS.init_pop_full — Method

Function that initializes a population of size popSize using the full method.

init_pop_full(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Every initialization functions take the same parameters, but not all do of them makes use of all parameters. This is just to unify the call of initialization functions.

GP_NLS.init_pop_grow — Method

Function that initializes a population of size popSize using the grow method.

init_pop_grow(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Every initialization functions take the same parameters, but not all do of them makes use of all parameters. This is just to unify the call of initialization functions.

GP_NLS.init_pop_ramped — Method

Function that initializes a population of size popSize using the ramped half-half method.

init_pop_ramped(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)

Every initialization functions take the same parameters, but not all do of them makes use of all parameters. This is just to unify the call of initialization functions.