Conteúdo dos nós das árvores

Toda a representação por árvores começa aqui. Essas Structs são utilizadas como conteúdo dos nós da árvore (e não como nós diretos).

Tipos e Funções

GP_NLS.defaultConstSet — Constant

Conjunto de constantes padrão

Const(3.1415),
Const(1.0),
Const(-1.0)

GP_NLS.defaultERCSet — Constant

Conjunto de ERC padrão

ERC(-1.0, 1.0)

GP_NLS.defaultFunctionSet — Constant

Conjunto de funções padrão

Func(+, 2),
Func(-, 2),
Func(myprod, 2),
Func(mydiv, 2),

Func(mysquare, 1),
Func(mysqrt, 1),
Func(myexp, 1),
Func(mylog, 1)

GP_NLS.Const — Type

Struct para representar um valor Float64 constante nos terminais das árvores.

Const(v::Float64)

Recebe um valor Float64 v que será utilizado como constante.

A representação da constante como string é obtida arredondando o valor para 3 casas decimais, e é obtida automaticamente

O método de otimização com mínimos quadrados não linear busca por essa struct especificamente para fazer a otimização dos seus valores.

GP_NLS.ERC — Type

Struct que guarda os limites do intervalo de criação de uma constante aleatória.

ERC(lb::Float64, ub::Float64)

Essa struct é utilizada para criação de constantes nos terminais quando é selecionada, sendo que será criado um terminal com a struct Const com um valor aleatório sorteado entre [lb, ub) para ocupar o lugar do ERC (Ephemeral Random Constant) no terminal. A representação em string da constante criada é como descrito na documentação de Const.

GP_NLS.Func — Type

Struct para representar uma função de um nó interno de uma árvore de expressão.

Func(f::Function, a::Int64)

Recebe uma função f e a aridade da função a. A função deve sempre funcionar na forma vetorizada (receberá sempre $a$ arrays (com $n$ valores, onde cada um é uma observação, sendo cada array um argumento da função) e é aplicada sobre os arrays. Então a entrada da função terá a linhas e n colunas.

A representação da função em string é inferida do nome da função passada, e quando essa função for utilizada para criar um nó de uma árvore, ela terá $a$ filhos.

Na implementação dessa biblioteca, não devem ser utilizados operadores protegidos, já que é feito o uso de um autodiff para diferenciar a árvore, e pode ser problemático diferenciar funções com IFs ou cálculos mais elaborados para proteger a função.

GP_NLS.Var — Type

Struct que representa uma variável do problema, e é utilizada nos terminais das árvores.

Var(v::String, i::Int64)

Recebe uma string v que será utilizada como representação da variável na hora de imprimir os dados (pode-se utilizar um placeholder caso a base de dados não tenha nome nas colunas) e um índice ì que corresponde ao índice da coluna da variável correspondente na matriz da base de dados.

GP_NLS.WeightedVar — Type

Struct de uma variável com coeficiente. Utilizar quando quisermos que as variáveis criadas tenham sempre um coeficiente associado à elas no momento de criação.

WeightedVar(v::String, i::Int64) = new(v, i, v)

Essa struct representa uma variável com peso que pode ser ajustado pelo método de otimização (ou definido antes da execução pelo utilizador, por meio do múltiplo despache do construtor da struct).

A representação em String é inferida da forma que uma $Var$ em relação ao nome da variável, e o coeficiente é inferido da mesma forma que uma $Const$.

Essa variável ponderada é, na teoria, uma sub-árvore com 3 nós e profundidade 2, mas na prática é tratada como um único nó pois não é de interesse fazer uma dissociação entre o peso e a variável durante o processo evolutido. Ao tratar a variável com peso como um único nó, não é necessario modificar nenhuma implementação de crossover ou mutação para que não seja destruída essa subárvore.

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children