Funções de utilidade

Implementação de algumas funções auxiliares que são utilizadas para inspecionar e manipular as árvores de forma mais geral.

Tipos e Funções

GP_NLS.branches_in_limits — Function

Encontra todos os galhos de uma árvore qualquer node que tem uma quantidade de nós menor ou igual que allowedSize e uma profundidade menor ou igual que allowedDepth . Retorna uma lista com a posição de todos os galhos encontrados.

branches_in_limits(
    allowedSize::Int64, allowedDepth::Int64, node::AbstractNode, _point::Int64=1)::Vector{Int64}

O parâmetro _point é de uso interno, e serve para monitorar o ponto da árvore onde foram encontrados os candidatos. Em chamadas recursivas ele faz sentido, mas fora da função ele não representa nenhuma informação útil.

GP_NLS.change_at! — Method

Recebe um ponto p do tipo inteiro (que deve ser menor ou igual ao número de nós da árvore passada), um galho do tipo AbstractNode e um nó qualquer representando uma árvoreAbstractNode, e insere o galho na árvore passada no nó de posição p. Altera a árvore passada como argumento.

change_at!(p::Int64, branch::AbstractNode, node::AbstractNode)::AbstractNode

Recebe um ponto, uma subárvore para inserir, e a árvore onde será inserida. Esse método que realiza a modificação de uma árvore no crossover.

GP_NLS.change_children! — Method

Função que recebe um ponto p (que deve ser menor ou igual ao número de nós da árvore passada), um galho do tipo AbstractNode e uma lista de filhos Vector{AbstractNode} e retorna uma modificação da lista de filhos, sendo que o nó de posição p será substituído pelo galho passado. Altera a lista de filhos passada como argumento.

change_children!(p::Int64, branch::AbstractNode, children::Vector{AbstractNode})::Vector{AbstractNode}

Esse método é auxiliar de change_at! e de uso interno da biblioteca.

GP_NLS.copy_tree — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente recria a estrutura, de forma que as referências não sejam compartilhadas entre a árvore passada e a árvore retornada, evitando efeitos colaterais no manuseio. Implementa despache múltiplo.

copy_tree(node::AbstractNode)::TerminalNode

GP_NLS.depth — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente encontra a profundidade da árvore como sendo o tamanho de seu maior galho. Implementa múltiplo despache.

depth(node::AbstractNode)::Int64

Essa profundidade corresponde ao número de nós existentes, sendo que os nós de variáveis ponderadas (que são uma subárvore com profundidade 2) ainda são considerados com profundidade 1.

GP_NLS.get_branch_at — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e um inteiro p (que deve ser menor ou igual ao número de nós da árvore passada) e retorna o galho na posição p.

get_branch_at(p::Int64, node::AbstractNode)::AbstractNode

GP_NLS.get_depth_at — Method

Recebe um ponto p do tipo inteiro (que deve ser menor ou igual ao número de nós da árvore passada) e um nó qualquer representando uma árvoreAbstractNode, e encontra a profundidade da subárvore na posição p.

get_depth_at(p::Int64, node::AbstractNode)::Int64

É como o método de obter a profundidade, mas esse método calcula a profundidade parcial de uma subárvore que está no ponto p.

GP_NLS.getstring — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente constrói uma string representando a árvore passada. Funções são denotadas sempre na notação prefixada, com os argumentos entre parênteses.

getstring(node::AbstractNode)::String

Implementa um despache múltiplo para cada subtipo de AbstractNode.

GP_NLS.numberofnodes — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente conta o total de nós que a árvore possui. Implementa múltiplo despache.

numberofnodes(node::AbstractNode)::Int64

Essa função não considera variáveis ponderadas como um único nó.

GP_NLS.true_depth — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente encontra a profundidade da árvore como sendo o tamanho de seu maior galho. Implementa múltiplo despache.

depth(node::AbstractNode)::Int64

Essa profundidade é real, e corresponde ao número de nós existentes, considerando a profundidade da subárvore de variáveis com pesos. Essa função não é utilizada nas implementações, e fica disponível para o usuário caso deseje obter a profundidade real considerando variáveis ponderadas.

GP_NLS.true_numberofnodes — Method

Função que recebe um nó qualquer de uma árvore (AbstractNode) e recursivamente conta o total de nós que a árvore possui. Implementa múltiplo despache.

true_numberofnodes(node::AbstractNode)::Int64

Essa função não considera variáveis ponderadas como três nós. Essa função só é utilizada nas operações de mutação, crossover e inicialização para evitar árvores maiores que o permitido.

GP_NLS.which_children — Method

Função que recebe o conjunto de filhos children::Vector{AbstractNode} e um inteiro p (que deve ser menor ou igual ao número de nós somados de todos os filhos passados) e encontra o filho que contém o p-ésimo nó da árvore caso fosse percorrida na ordem nó > filho 1 > filho 2 > .... p deve ser obtida pelo número de nós da árvore (não o número de nós real, considerando vários nós em variáveis com peso).

which_children(p::Int64, children::Vector{AbstractNode})::Tuple{Int64, AbstractNode}

Retorna uma tupla contendo o filho que contém o nó de índice desejado, e um inteiro informando a posição do p-ésimo nó dentro da sub-árvore (filho) retornada.

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children