Default sets

There are some default sets available if the user does not want to understand how to create new constants and functions.

Const sets

GP_NLS.defaultConstSet — Constant

Default const set

Const(3.1415),
Const(1.0),
Const(-1.0)

GP_NLS.defaultERCSet — Constant

Default ERC set

ERC(-1.0, 1.0)

GP_NLS.defaultFunctionSet — Constant

Default functions set

Func(+, 2),
Func(-, 2),
Func(prod, 2),
Func(div, 2),

Func(square, 1),
Func(sqrt, 1),
Func(exp, 1),
Func(log, 1)