Algoritmo evolutivo

Implementação da mutação, crossover e GP.

Tipos e Funções

GP_NLS.GP — Function

GP Com controle de profundidade e número de nós. A inicialização recomendada é a PTC2, mas temos as outras também (entretanto, os outros métodos são baseados no GP do koza e não seguem restrição de número máximo de nós).

GP(
    X::Matrix{Float64}, 
    y::Vector{Float64},
    fSet::Vector{Func},
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
    minDepth::Int64        = 1,
    maxDepth::Int64        = 5,
    maxSize::Int64         = 25,
    popSize::Int64         = 50,
    gens::Int64            = 50,
    mutationRate::Float64  = 0.25,
    elitism::Bool          = false,
    verbose::Bool          = false,
    init_method::String    = "PTC2", #["ramped", "grow", "full", "PTC2"]
    lm_optimization        = false, 
    keep_linear_transf_box = false
)::AbstractNode

GP_NLS.crossover — Method

Função de crossover que faz uma recombinação dos dois pais passados como argumento, encontrando um ponto de quebra em cada uma das árvores pais e trocando a subárvore entre esses pontos. Não modifica os pais. Esse crossover controla o número de nós (e não a profundidade).

crossover(
    fst_parent::AbstractNode,
    snd_parent::AbstractNode,
    maxDepth::Int64,
    maxSize::Int64)::AbstractNode

Queremos que o crossover ocorra em um ponto de corte que não causa o aumento da árvore filha para algo maior que o limite.

GP_NLS.mutation! — Method

Função que implementa uma mutação tradicional de substituição em uma árvore, respeitando a profundidade máxima passada. Modifica a árvore passada.

mutation!(
    node::AbstractNode,
    maxDepth::Int64,
    maxSize::Int64,
    fSet::Vector{Func},
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}},
    mutationRate::Float64
)::AbstractNode

A taxa de mutação mutationRate deve variar em $[0, 1]$ e determina a chance de ocorrer uma mutação (substituindo um ponto aleatório por uma nova sub-árvore) ou o próprio nó passado ser retornado.

GP_NLS.tourn_selection — Method

Função que recebe dois indivíduos e faz uma seleção por torneio simples. Um indivíduo é uma tupla (fitness::Float64, node::AbstractNode) com a árvore e seu fitness, para evitar que seja recalculado a todo instante.

tourn_selection(ind1::Tuple{Float64, AbstractNode}, ind2::Tuple{Float64, AbstractNode})::Tuple{AbstractNode, Float64}

O retorno é uma tupla com o indivíduo vencedor (ou seja, é retornada uma tupla).

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children