Inicialização da população

Implementação de diferentes funções para criar árvores, e funções que criam uma população inteira

Tipos e Funções

GP_NLS.PTC2 — Method

Criação de árvores com o método Probabilistic Tree Creator 2 (PTC2), descrito em Two Fast Tree-Creation Algorithms for Genetic Programming, de Sean Luke.

Esse método se parece com o método full de Koza, mas além de respeitar um limite de profundidade maxDepth, respeita um limite de quantidade de nós expctdSize. O PTC2 garante que a profundidade não ultrapasse o máximo (no nosso caso, variáveis com peso contam como profundidade 1), e garante que o número de nós não ultrapasse o valor esperado somado da maior aridade entre as funções, isso é, $expctdSize + max(aridade(f)), f in fSet$.

PTC2(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    maxDepth::Int64,
    expctdSize::Int64)::AbstractNode

Aqui adotamos que a chance de selecionar um terminal $t$ quando for necessário inserir um terminal será uniforme para todos os terminais, e a chance de inserir uma função também seguirá a mesma lógica.

O algoritmo do PTC2 é descrito em C e faz o uso de pilhas e ponteiros. Em Julia, não há todos esses recursos de forma simples, e uma adaptação dessas funções foi feita.

GP_NLS._create_random_terminal — Method

Função que recebe o conjunto de conteúdos terminais possíveis e faz a criação de um nó terminal. Essa função, de uso interno, serve para criar um nó terminal com o tratamento do ERC.

_create_random_terminal(tSet::Vector{Union{Var, WeightedVar, Const, ERC}})::TerminalNode

A criação de um nó terminal envolve um passo de verificação adicional para o caso de ERC, que deve substituir o nó por uma constante dentro do intervalo especificado.

GP_NLS.full — Method

Função que cria uma árvore pelo método full, inspirado no trabalho original de Koza. Recebe um conjunto de funções fSet::Vector{Func} que serão sorteadas para os nós internos, um conjunto de funções tSet::Vector{Union{Var, WeightedVar, Const, ERC}} que serão utilizadas nos terminais, e uma profundidade máxima maxDepth::Int64 permitida. Retorna uma árvore qualquer com profundidade máxima maxDepth e utilizando os conteúdos de funções e terminais passados.

full(fSet::Vector{Func}, tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, maxDepth::Int64)::AbstractNode

GP_NLS.grow — Method

Função que cria uma árvore pelo método grow, inspirado no trabalho original de Koza. Recebe um conjunto de funções fSet::Vector{Func} que serão sorteadas para os nós internos, um conjunto de funções tSet::Vector{Union{Var, WeightedVar, Const, ERC}} que serão utilizadas nos terminais, e uma profundidade máxima maxDepth::Int64 permitida. Retorna uma árvore qualquer com profundidade máxima maxDepth e utilizando os conteúdos de funções e terminais passados.

grow(fSet::Vector{Func}, tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, maxDepth::Int64)::AbstractNode

Repare que não há tamanho mínimo, significando que pode ser retornada uma árvore de um único nó. A profundidade máxima considera variáveis com peso como um único nó.

GP_NLS.init_pop_PTC2 — Method

Função que inicializa uma população de tamanho popSize utilizando o método PTC2.

init_pop_PTC2(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Todas as funções de inicialização recebem os mesmos parâmetros, mas nem todas fazem uso de todos eles. Isso é apenas para unificar a chamada da criação de populações

GP_NLS.init_pop_full — Method

Função que inicializa uma população de tamanho popSize utilizando o método full.

init_pop_full(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Todas as funções de inicialização recebem os mesmos parâmetros, mas nem todas fazem uso de todos eles. Isso é apenas para unificar a chamada da criação de populações

GP_NLS.init_pop_grow — Method

Função que inicializa uma população de tamanho popSize utilizando o método grow.

init_pop_grow(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)::Vector{AbstractNode}

Todas as funções de inicialização recebem os mesmos parâmetros, mas nem todas fazem uso de todos eles. Isso é apenas para unificar a chamada da criação de populações

GP_NLS.init_pop_ramped — Method

Função que inicializa uma população de tamanho popSize utilizando o método ramped half-half.

init_pop_ramped(
    fSet::Vector{Func}, 
    tSet::Vector{Union{Var, WeightedVar, Const, ERC}}, 
    minDepth::Int64,
    maxDepth::Int64,
    expctdSize::Int64,
    popSize::Int64)

Todas as funções de inicialização recebem os mesmos parâmetros, mas nem todas fazem uso de todos eles. Isso é apenas para unificar a chamada da criação de populações

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children