Estrutura das árvores

Para construir as árvores de expressões, são utilizadas as structs definidas aqui, que funcionam como um esqueleto das árvores.

Tipos e Funções

GP_NLS.AbstractNode — Type

Abstract type of our symbolic tree nodes. The idea of creating this type is to create subtypes InternalNode and TerminalNode, and then use multiple dispatch to implement functions that should have different behavior when manipulating the expression trees.

Expression trees should not be built using Var, WeightedVar, Const, Func, ERC, but with these nodes, which are intended to be used as the "backbone" of the tree. The backbone is build using the InternalNode and TerminalNode, and its contents should be the ones declared in $NodeContent.jl$.

A terminal node must have as its contents only the types Const, Var, WeightedVar, and an internal nodemust have its contents the type Func.

Notice that the ERC, when selected to be used in a terminal during the creation of a tree, is replaced by a random Const node). There are not any explicit ERC terminal in the trees.

GP_NLS.InternalNode — Type

Struct to build the internal nodes of the backbone of the tree. Its contents will always be of the func type, a f ::Func function that will necessarily have f.arity children, where f.arity is the arity of the function.

InternalNode(f::Func, children::Vector{AbstractNode}) <: AbstractNode

GP_NLS.TerminalNode — Type

Struct to build the terminal nodes of the backbone of the tree. Its contents will always be of the terminal type: Union{Const, Var, WeightedVar}.

TerminalNode(terminal::Union{Const, Var, WeightedVar}) <: AbstractNode

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children