Funções de utilidade

Implementação de algumas funções auxiliares que são utilizadas para inspecionar e manipular as árvores de forma mais geral.

Tipos e Funções

GP_NLS.branches_in_limits — Function

Finds all branches of any tree node that have a number of nodes less than or equal to allowedSize and a depth less than or equal to allowedDepth . Returns a list with the position of all branches found.

branches_in_limits(
    allowedSize::Int64, allowedDepth::Int64, node::AbstractNode, _point::Int64=1)::Vector{Int64}

The _point parameter being returned is for internal use, and serves to monitor the point of the tree where the candidates were found. This is only meaningful to recursive calls, and outside of the function it does not represent any useful information.

GP_NLS.change_at! — Method

Takes a point p of type integer (which must be less than or equal to the number of nodes of branch), a branch of type AbstractNode, and any node AbstractNode representing a tree. Returns a modification of the tree by inserting the branch into the tree at position `$p$. This function changes the tree passed as argument.

change_at!(p::Int64, branch::AbstractNode, node::AbstractNode)::AbstractNode

This method is mainly used to modify trees in the crossover operation.

GP_NLS.change_children! — Method

Function that takes a point p (which must be less than or equal to the number of nodes of branch), a branch of type AbstractNode, and a list of children Vector{AbstractNode}. Returns a modification of the list of children, where the p position node will be replaced by the given branch. This function changes the list of children passed as argument.

change_children!(p::Int64, branch::AbstractNode, children::Vector{AbstractNode})::Vector{AbstractNode}

This function is a helper to change_at!, and itsfor internal use. It is not exported by the module.

GP_NLS.copy_tree — Method

Function that takes any node of a tree (AbstractNode) and recursively recreates the structure so that references are not shared between the tree passed and the tree returned, avoiding side effects when manipulating the tree.

copy_tree(node::AbstractNode)::TerminalNode

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.depth — Method

Function that takes any node of a tree (AbstractNode) and recursively finds the depth of the tree, where the depth is the size of its largest branch. This depth function does not take into count the coefficients of weighted variables (which are, in fact, a subtree with depth 2). To find the depth of a tree considering weighted variables as a subtree (and not as a single node), use the function true_depth.

depth(node::AbstractNode)::Int64

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.get_branch_at — Method

Function that takes any node of a tree (AbstractNode) and an integer p (which must be less than or equal to the number of nodes in tree) and returns the branch at position p.

get_branch_at(p::Int64, node::AbstractNode)::AbstractNode

GP_NLS.get_depth_at — Method

Takes a point p of type integer (which must be less than or equal to the number of nodes of the node passed) and any node AbstractNode representing a tree, then this function find and return the depth of the subtree at position p.

get_depth_at(p::Int64, node::AbstractNode)::Int64

It's almost like finding the depth of the tree, but when we are interested in finding the depth of a subtree that is at the point p, not the whole tree depth.

GP_NLS.getstring — Method

Function that takes any node of a tree (AbstractNode) and recursively builds a string representation of the tree, where functions are always denoted in prefixed notation, with arguments in parentheses.

getstring(node::AbstractNode)::String

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.numberofnodes — Method

Function that takes any node of a tree (AbstractNode) and recursively counts the total number of nodes of the tree. This function counts weighted variables as a single node. To find the number of nodes of a tree considering weighted variables as a subtree (and not as a single node), use the function true_numberofnodes.

numberofnodes(node::AbstractNode)::Int64

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.true_depth — Method

Function that takes any node of a tree (AbstractNode) and recursively finds the depth of the tree, where the depth is the size of its largest branch. This depth function returns a value that corresponds to the number of existing nodes in the tree, considering weighted variables as being a subtree of depth 2. This function is not used in implementations, and is available to users who want to get the real depth of GP-NLS trees.

true_depth(node::AbstractNode)::Int64

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.true_numberofnodes — Method

Function that takes any node of a tree (AbstractNode) and recursively counts the total number of nodes of the tree. This function counts weighted variables as three nodes. This function is only used in mutate, crossover, and initialize operations to avoid creating trees larger than the allowed.

true_numberofnodes(node::AbstractNode)::Int64

This function works by having a multiple dispatch for each subtype of AbstractNode.

GP_NLS.which_children — Method

Function that takes a set of children of the same node as the array children::Vector{AbstractNode}, and an integer p (which must be less than or equal to number of nodes in children) and finds the child containing the p-th node of the children if it was traversed inorder.

Returns a tuple containing the child which contains the node of index p, and an integer informing the position of the p-th node within the returned (child) subtree.

which_children(p::Int64, children::Vector{AbstractNode})::Tuple{Int64, AbstractNode}

Índice

GP_NLS.defaultConstSet

GP_NLS.defaultERCSet

GP_NLS.defaultFunctionSet

GP_NLS.AbstractNode

GP_NLS.Const

GP_NLS.ERC

GP_NLS.Func

GP_NLS.InternalNode

GP_NLS.TerminalNode

GP_NLS.Var

GP_NLS.WeightedVar

GP_NLS.GP

GP_NLS.PTC2

GP_NLS._create_random_terminal

GP_NLS.branches_in_limits

GP_NLS.change_at!

GP_NLS.change_children!

GP_NLS.copy_tree

GP_NLS.crossover

GP_NLS.depth

GP_NLS.evaluate

GP_NLS.fitness

GP_NLS.full

GP_NLS.get_branch_at

GP_NLS.get_depth_at

GP_NLS.getstring

GP_NLS.grow

GP_NLS.init_pop_PTC2

GP_NLS.init_pop_full

GP_NLS.init_pop_grow

GP_NLS.init_pop_ramped

GP_NLS.mutation!

GP_NLS.numberofnodes

GP_NLS.tourn_selection

GP_NLS.true_depth

GP_NLS.true_numberofnodes

GP_NLS.which_children